+++ tests/Fake/FactoryWithInjectedProperty.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Fake;
+
+use Gacela\Container\Attribute\Factory;
+use Gacela\Container\Attribute\Inject;
+
+#[Factory]
+final class FactoryWithInjectedProperty
+{
+    #[Inject]
+    public ClassWithoutDependencies $dependency;
+}